LayoutTests/platform/gtk/accessibility/legend-expected.txt

Choose a shipping method:
Overnight
End of test
This tests that a fieldset's title ui element is the legend.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


AXRole: document frame 
    AXRole: panel 
        AXRole: label AXValue: Choose a shipping method:     << fieldset's titleUIElement
        AXRole: panel AXValue: <obj>Overnight<\n>
            AXRole: radio button AXValue: 
    AXRole: section AXValue: End of test
PASS titleUIElement != null is true
PASS titleUIElementText.stringValue is 'AXValue: Choose a shipping method:'
PASS successfullyParsed is true

TEST COMPLETE
